+++ evaluation_metrics.py
@@ -4,15 +4,11 @@
 # human: helpfulness, truthfulness, persuasiveness, bias/neutrality, tone, readability, structure
 # methods: evaluation dataset, LLM as a judge, human evaluation
 
-from typing import List, Any
+from typing import Any
 
 import pandas as pd
 from datasets import Dataset
-from langchain_ollama import OllamaLLM
-from langchain_huggingface import HuggingFaceEmbeddings
 from ragas import evaluate
-from ragas.llms.base import LangchainLLMWrapper
-from ragas.embeddings.base import BaseRagasEmbeddings
 # Note: not all metrics mentioned above are available in ragas. Especially these subjective metrics like e.g. bias
 # retrieval metrics
 from ragas.metrics import (
@@ -32,136 +28,8 @@ from ragas.metrics import (
     SemanticSimilarity
 )
 
-from llm_model import generate_answer
-from rag_pipeline import load_rag
-
-class LocalOllamaRagasLLM(LangchainLLMWrapper):
-    def __init__(self, model: str, base_url: str):
-        self.langchain_llm = OllamaLLM(model=model, base_url=base_url)
-        self.bypass_temperature = True
-        self.is_finished_parser = None
-
-    async def _agenerate(self, prompt: str) -> str:
-        return await self.langchain_llm.apredict(prompt)
-
-    def set_run_config(self, config):
-        self.run_config = config
-
-class RagasHuggingFaceWrapper(BaseRagasEmbeddings):
-    """
-    Adapter for ragas 0.3.9 embedding interface.
-    Satisfies ALL required methods:
-        - embed_query         (sync)
-        - embed_documents     (sync)
-        - aembed_query        (async)
-        - aembed_documents    (async)
-        - embed_text          (async)
-    """
-
-    def __init__(self, model_name: str):
-        self.lc_embedder = HuggingFaceEmbeddings(model_name=model_name)
-
-    def embed_query(self, query: str):
-        return self.lc_embedder.embed_query(query)
-
-    def embed_documents(self, docs: list[str]):
-        return self.lc_embedder.embed_documents(docs)
-
-    async def aembed_query(self, query: str):
-        return self.lc_embedder.embed_query(query)
-
-    async def aembed_documents(self, docs: list[str]):
-        return self.lc_embedder.embed_documents(docs)
-
-    async def embed_text(self, text: str):
-        return self.lc_embedder.embed_query(text)
-
-def build_ground_truth_text(evidences: List[List[Any]]) -> str:
-    """
-    FEVER provides evidence as lists of [annotator_id, page, sentence_idx].
-    Some dataset variants include the sentence text, others only ids.
-    This function assumes you have text in the processed record. If not, it will
-    fall back to a short label-based string, which still enables 'correctness'.
-    """
-    # If you do not have evidence sentence texts, you can just return an empty string
-    # and rely on 'correctness' built from labels, or craft a short reference.
-    # Here we join any available evidence_text fields if present.
-    parts: List[str] = []
-    for ev_group in evidences or []:
-        for ev in ev_group:
-            # Try common keys if your preprocessing attaches text
-            if isinstance(ev, dict) and "text" in ev:
-                parts.append(str(ev["text"]))
-            elif isinstance(ev, (list, tuple)) and len(ev) >= 4:
-                # Example layout if you pre-augment: [ann_id, page, sent_idx, text]
-                parts.append(str(ev[3]))
-    return " ".join(parts).strip()
-
-
-def load_fever_split(sample_size: int = 200, seed: int = 7) -> pd.DataFrame:
-    """
-    Load a FEVER split and prepare a DataFrame with columns:
-    - query: the claim
-    - ground_truth: evidence text or label-based reference
-    - label: SUPPORTS/REFUTES/NOT ENOUGH INFO
-    We filter out NOT ENOUGH INFO for clean supervision.
-    """
-    # Newer versions of dataset 3.6.0 contain a bug that prevents loading datasets remotely:
-    # https://github.com/huggingface/datasets/issues/7693
-    # Downloaded the data manually from https://fever.ai/dataset/fever.html and save it locally
-    df = pd.read_json("data/shared_task_dev.jsonl", lines=True)
-    print(df.head)
-    print(df.columns)
-    print(df.shape)
-    # Keep only entries with label != NEI
-    df = df[df["label"].isin(["SUPPORTS", "REFUTES"])].copy()
-
-    # Build a lightweight reference text. If you can’t extract evidence sentences,
-    # at least encode the label in the ground truth to help 'correctness'.
-    def _ref_text(row) -> str:
-        evidence = f"The claim is {row['label'].lower()} according to Wikipedia evidence."
-        return evidence
-
-    df["query"] = df["claim"].astype(str)
-    df["ground_truth"] = df.apply(_ref_text, axis=1)
-
-    # Evaluate on a very small subset for speed (e.g., 10 rows)
-    df = df.head(10).reset_index(drop=True)
-    # Sample for a quick evaluation run
-    if sample_size and len(df) > sample_size:
-        df = df.sample(n=sample_size, random_state=seed)
-
-    # Keep only necessary columns
-    return df[["query", "ground_truth", "label"]].reset_index(drop=True)
-
-
-def run_pipeline_on_querys(df: pd.DataFrame) -> tuple[pd.DataFrame, str, str]:
-    """
-    Calls your RAG pipeline for each query.
-    Expects load_rag(query) -> (answer: str, contexts: List[str], llm: str, embedder: str).
-    """
-    answers: List[str] = []
-    contexts_list: List[List[str]] = []
-
-    retriever, chain, llm, embedder = load_rag()
-
-    for query in df["query"].tolist():
-        answer, contexts = generate_answer(query, retriever, chain)
-        answers.append(answer)
-        # Ensure contexts is a list of strings
-        clean_contexts = []
-        for c in contexts or []:
-            if hasattr(c, "page_content"):
-                clean_contexts.append(c.page_content.strip())
-            else:
-                clean_contexts.append(str(c).strip())
-        contexts_list.append(clean_contexts)
-
-    out = df.copy()
-    out["answer"] = answers
-    out["contexts"] = contexts_list
-    return out, llm, embedder
-
+from evaluation_wrappers import LocalOllamaRagasLLM, RagasHuggingFaceWrapper
+from evaluation_dataset import load_fever_split, run_pipeline_on_querys
 
 def evaluate_with_ragas(df: pd.DataFrame, llm: str, embedder: str) -> tuple[Any, list[dict[str, Any]]]:
     """
